chore: cleanup transaction endpoints
